# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users/me": {
      "get": {
        "description": "Devuelve el perfil del personal autenticado, sin storeName ni contraseña. Roles permitidos: OWNER, MANAGER, CASHIER.",
        "operationId": "UsersController_getMyDataUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación completada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffUser"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, propiedades no permitidas o regla de negocio incumplida. En usuarios, también se usa cuando el usuario no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "JWT ausente, inválido o vencido; usuario inexistente o inactivo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno o de base de datos. Algunos errores Prisma no interceptados se devuelven con este estado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Consultar mi perfil",
        "tags": [
          "users"
        ],
        "x-roles": [
          "OWNER",
          "MANAGER",
          "CASHIER"
        ]
      },
      "patch": {
        "description": "Actualiza firstName, lastName, email, phone y role. El DTO actual también acepta password, pero el servicio lo ignora; use me/password. Actualmente también permite enviar role en esta ruta. Roles permitidos: OWNER, MANAGER, CASHIER.",
        "operationId": "UsersController_updateMyData",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operación completada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffUser"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, propiedades no permitidas o regla de negocio incumplida. En usuarios, también se usa cuando el usuario no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "JWT ausente, inválido o vencido; usuario inexistente o inactivo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno o de base de datos. Algunos errores Prisma no interceptados se devuelven con este estado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar mi perfil",
        "tags": [
          "users"
        ],
        "x-roles": [
          "OWNER",
          "MANAGER",
          "CASHIER"
        ]
      }
    },
    "/api/users/me/password": {
      "patch": {
        "description": "Comprueba la contraseña actual y exige una nueva diferente de al menos seis caracteres. Devuelve el perfil del personal. Roles permitidos: OWNER, MANAGER, CASHIER.",
        "operationId": "UsersController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operación completada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffUser"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, propiedades no permitidas o regla de negocio incumplida. En usuarios, también se usa cuando el usuario no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "JWT ausente, inválido o vencido; usuario inexistente o inactivo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno o de base de datos. Algunos errores Prisma no interceptados se devuelven con este estado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cambiar mi contraseña",
        "tags": [
          "users"
        ],
        "x-roles": [
          "OWNER",
          "MANAGER",
          "CASHIER"
        ]
      }
    },
    "/api/users": {
      "post": {
        "description": "Crea un registro asociado a la tienda del usuario autenticado. Roles permitidos: OWNER, MANAGER.",
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Operación completada (POST devuelve 201).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffUser"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, propiedades no permitidas o regla de negocio incumplida. En usuarios, también se usa cuando el usuario no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "JWT ausente, inválido o vencido; usuario inexistente o inactivo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Requiere uno de estos roles: OWNER, MANAGER.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno o de base de datos. Algunos errores Prisma no interceptados se devuelven con este estado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear usuario",
        "tags": [
          "users"
        ],
        "x-roles": [
          "OWNER",
          "MANAGER"
        ]
      },
      "get": {
        "description": "Listado paginado por tienda. Busca por nombre, apellido, correo o teléfono. Orden predeterminado createdAt desc. Roles permitidos: OWNER, MANAGER.",
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Página solicitada; enviar un entero positivo",
            "schema": {
              "minimum": 0,
              "exclusiveMinimum": true,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Elementos por página; enviar un entero positivo",
            "schema": {
              "minimum": 0,
              "exclusiveMinimum": true,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Texto de búsqueda, sin distinguir mayúsculas",
            "schema": {
              "example": "cafe",
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Campo del modelo Prisma para ordenar; no usar nombres calculados de la respuesta",
            "schema": {
              "default": "createdAt",
              "example": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "direction",
            "required": false,
            "in": "query",
            "description": "Dirección del orden",
            "schema": {
              "default": "desc",
              "example": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación completada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffUserPage"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, propiedades no permitidas o regla de negocio incumplida. En usuarios, también se usa cuando el usuario no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "JWT ausente, inválido o vencido; usuario inexistente o inactivo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Requiere uno de estos roles: OWNER, MANAGER.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno o de base de datos. Algunos errores Prisma no interceptados se devuelven con este estado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar usuarios",
        "tags": [
          "users"
        ],
        "x-roles": [
          "OWNER",
          "MANAGER"
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "description": "Obtiene el registro identificado por UUID dentro de la tienda del usuario. Usuario inexistente devuelve 400. Roles permitidos: OWNER, MANAGER.",
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del recurso en la tienda del usuario.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación completada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffUser"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, propiedades no permitidas o regla de negocio incumplida. En usuarios, también se usa cuando el usuario no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "JWT ausente, inválido o vencido; usuario inexistente o inactivo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Requiere uno de estos roles: OWNER, MANAGER.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno o de base de datos. Algunos errores Prisma no interceptados se devuelven con este estado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Consultar usuario",
        "tags": [
          "users"
        ],
        "x-roles": [
          "OWNER",
          "MANAGER"
        ]
      },
      "patch": {
        "description": "Actualiza datos y rol. password es aceptado por el DTO pero ignorado por este servicio. Roles permitidos: OWNER, MANAGER.",
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del recurso en la tienda del usuario.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operación completada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffUser"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, propiedades no permitidas o regla de negocio incumplida. En usuarios, también se usa cuando el usuario no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "JWT ausente, inválido o vencido; usuario inexistente o inactivo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Requiere uno de estos roles: OWNER, MANAGER.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto de unicidad: el valor ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno o de base de datos. Algunos errores Prisma no interceptados se devuelven con este estado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar usuario",
        "tags": [
          "users"
        ],
        "x-roles": [
          "OWNER",
          "MANAGER"
        ]
      },
      "delete": {
        "description": "Elimina el registro y devuelve sus datos. Las relaciones existentes pueden impedir la eliminación. Roles permitidos: OWNER.",
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del recurso en la tienda del usuario.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación completada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffUser"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, propiedades no permitidas o regla de negocio incumplida. En usuarios, también se usa cuando el usuario no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "JWT ausente, inválido o vencido; usuario inexistente o inactivo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Requiere uno de estos roles: OWNER.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto de unicidad: el valor ya existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno o de base de datos. Algunos errores Prisma no interceptados se devuelven con este estado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar usuario",
        "tags": [
          "users"
        ],
        "x-roles": [
          "OWNER"
        ]
      }
    },
    "/api/users/{id}/status": {
      "patch": {
        "description": "Invierte isActive; no recibe cuerpo. No fija un estado explícito. Roles permitidos: OWNER.",
        "operationId": "UsersController_changeActiveStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del recurso en la tienda del usuario.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación completada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffUser"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, propiedades no permitidas o regla de negocio incumplida. En usuarios, también se usa cuando el usuario no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "JWT ausente, inválido o vencido; usuario inexistente o inactivo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Requiere uno de estos roles: OWNER.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno o de base de datos. Algunos errores Prisma no interceptados se devuelven con este estado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Alternar estado activo",
        "tags": [
          "users"
        ],
        "x-roles": [
          "OWNER"
        ]
      }
    }
  },
  "info": {
    "title": "La Hermandad — API POS",
    "description": "API del punto de venta: autenticación, usuarios, categorías, productos, clientes y pedidos con pagos. Las rutas protegidas obtienen la tienda desde el usuario autenticado. Use login o register y pegue el token en Authorize. Los JWT duran 24 horas. Los cuerpos rechazan propiedades desconocidas. Los importes de entrada son números y los Decimal de salida son cadenas. Todos los POST devuelven 201; PATCH y DELETE devuelven 200 con cuerpo. Los permisos se indican mediante x-roles en cada operación.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Pegue únicamente el token obtenido en auth/login o auth/register."
      }
    },
    "schemas": {
...
      "Role": {
        "type": "string",
        "enum": [
          "OWNER",
          "MANAGER",
          "CASHIER"
        ],
        "description": "Estado o rol permitido"
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "Nombre del usuario o cliente",
            "example": "Álvaro",
            "minLength": 3
          },
          "lastName": {
            "type": "string",
            "description": "Apellidos",
            "example": "Pérez",
            "minLength": 3
          },
          "email": {
            "type": "string",
            "description": "Correo electrónico",
            "example": "alvaro@example.com",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "description": "Teléfono en formato internacional",
            "example": "+59171234567"
          },
          "password": {
            "type": "string",
            "description": "Contraseña de acceso",
            "example": "Ejemplo123!",
            "format": "password",
            "writeOnly": true,
            "minLength": 6
          },
          "role": {
            "description": "Estado o rol permitido",
            "allOf": [
              {
                "$ref": "#/components/schemas/Role"
              }
            ]
          }
        }
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "description": "Contraseña actual",
            "example": "Ejemplo123!",
            "format": "password",
            "writeOnly": true
          },
          "newPassword": {
            "type": "string",
            "description": "Nueva contraseña, distinta de la actual",
            "example": "Nueva123!",
            "format": "password",
            "writeOnly": true,
            "minLength": 6
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "Nombre del usuario o cliente",
            "example": "Álvaro",
            "minLength": 3
          },
          "lastName": {
            "type": "string",
            "description": "Apellidos",
            "example": "Pérez",
            "minLength": 3
          },
          "email": {
            "type": "string",
            "description": "Correo electrónico",
            "example": "alvaro@example.com",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "description": "Teléfono en formato internacional",
            "example": "+59171234567"
          },
          "password": {
            "type": "string",
            "description": "Contraseña de acceso",
            "example": "Ejemplo123!",
            "format": "password",
            "writeOnly": true,
            "minLength": 6
          },
          "role": {
            "description": "Estado o rol permitido",
            "allOf": [
              {
                "$ref": "#/components/schemas/Role"
              }
            ]
          }
        },
        "required": [
          "firstName",
          "email",
          "password",
          "role"
        ]
      },
      "Error": {
        "type": "object",
        "required": [
          "statusCode",
          "message"
        ],
        "properties": {
          "statusCode": {
            "type": "integer",
            "example": 400
          },
          "message": {
            "description": "Mensaje del servicio o lista de errores de validación.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": [
              "property unknown should not exist"
            ]
          },
          "error": {
            "type": "string",
            "description": "Nombre HTTP del error, cuando Nest lo incluye",
            "example": "Bad Request"
          }
        }
      },
...
      "StaffUser": {
        "type": "object",
        "required": [
          "id",
          "firstName",
          "lastName",
          "email",
          "phone",
          "createdAt",
          "name",
          "initials",
          "avatar",
          "isShopOwner",
          "isActive",
          "role"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "123e4567-e89b-42d3-a456-426614174000"
          },
          "firstName": {
            "type": "string",
            "description": "Nombre",
            "example": "Álvaro"
          },
          "lastName": {
            "type": "string",
            "description": "Apellidos",
            "example": "Pérez",
            "nullable": true
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "alvaro@example.com",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "description": "Teléfono internacional",
            "example": "+59171234567",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-09-27T12:00:00.000Z"
          },
          "name": {
            "type": "string",
            "description": "Nombre completo",
            "example": "Álvaro Pérez"
          },
          "initials": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "Á",
              "P"
            ]
          },
          "avatar": {
            "type": "string",
            "description": "URL de avatar",
            "example": "Ejemplo",
            "nullable": true
          },
          "isShopOwner": {
            "type": "boolean",
            "example": true
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "MANAGER",
              "CASHIER"
            ],
            "example": "OWNER"
          }
        }
      },
...
    }
  }
}